Amend API URL, set cluster name via stringData
The use of the `_replace=true` query parameter caused me some problems
in testing (see https://github.com/rancher/rancher/issues/31534), so
remove this with the expectation that it's always set to a unique value.

Include the cluster name in the stringData instead of a seperate
environment variable to simplify what needs to be changed.